ARM: Stricten arguments to virt_to_pfn()
The kernel convention is that the argument passed to virt_to_pfn()
(virtual address to page frame number) needs to be an unsigned long,
but ARM has avoided this by including a cast to (unsigned long)
inside the definition of virt_to_pfn() itself. This creates problems
if we want to make virt_to_pfn() an actual function with a typed argument,
so fix up the offending call sites.

+++ arch/arm/include/asm/memory.h
@@ -392,9 +392,9 @@ static inline unsigned long __virt_to_idmap(unsigned long x)
  */
 #define ARCH_PFN_OFFSET		PHYS_PFN_OFFSET
 
-#define virt_to_page(kaddr)	pfn_to_page(virt_to_pfn(kaddr))
+#define virt_to_page(kaddr)	pfn_to_page(virt_to_pfn((unsigned long)kaddr))
 #define virt_addr_valid(kaddr)	(((unsigned long)(kaddr) >= PAGE_OFFSET && (unsigned long)(kaddr) < (unsigned long)high_memory) \
-					&& pfn_valid(virt_to_pfn(kaddr)))
+				 && pfn_valid(virt_to_pfn((unsigned long)(kaddr))))
 
 #endif
